# [jit] Adding vectorized load/store support for JIT generated CUDA kernel