@ngimel 2020-04-19
Can you please rebase and make sure that clang-format passes?