@FDecaYed 2020-04-20
> Can you please rebase and make sure that clang-format passes?

rebased and added commit for clang-format